backend/requirements.txt
gitpython
google-genai
python-dotenv
fastapi
uvicorn